refactor: move data exchange to a new tab
