+++ frontend/src/components/Modals/SettingsModal.vue
@@ -39,11 +39,12 @@
 </template>
 <script setup lang="ts">
 import { markRaw, ref } from 'vue'
-import { Code, Mailbox, TreePalm, User } from 'lucide-vue-next'
+import { Code, DatabaseBackup, Mailbox, TreePalm, User } from 'lucide-vue-next'
 import { Button, Dialog } from 'frappe-ui'
 
 import AccountSettings from '@/components/Settings/AccountSettings.vue'
 import AdvancedSettings from '@/components/Settings/AdvancedSettings.vue'
+import MailTransferSettings from '@/components/Settings/MailTransferSettings.vue'
 import ProfileSettings from '@/components/Settings/ProfileSettings.vue'
 import VacationResponseSettings from '@/components/Settings/VacationResponseSettings.vue'
 
@@ -64,6 +65,11 @@ const tabs = [
 		label: __('Vacation Response'),
 		icon: TreePalm,
 		component: markRaw(VacationResponseSettings),
+	},
+	{
+		label: __('Mail Transfer'),
+		icon: DatabaseBackup,
+		component: markRaw(MailTransferSettings),
 	},
 	{
 		label: __('Advanced'),
